final completo

+++ src/main/java/com/educantrol/educantrol_app/model/AsistenciaRepository.java
@@ -1,9 +1,10 @@
 package com.educantrol.educantrol_app.model;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AsistenciaRepository extends JpaRepository<Asistencia, Long> {
-
+public interface AsistenciaRepository extends JpaRepository<Asistencia, Integer> {
 }
+
